+++ samples/builtindb/get_builtindb.py
@@ -5,8 +5,8 @@ def get_data_builtindb(event, context):
     builtindb = BuiltInDB()
     collection_name = "sample-collection"
     data = event['data']
-    created_data = builtindb.database.insert(collection_name, data)
-    print(created_data)
-
-    get_data = builtindb.database.get(collection_name, created_data)
+    try:
+        get_data = builtindb.get(collection_name, data)
+    except Exception as e:
+        return str(e)
     return get_data